+++ src/Legemeldtsykefravarpanel/LegemeldtSykefravarPanel.less
