+++ src/Legemeldtsykefravarpanel/Lesmeriawebpanel/EkspanderbartInnhold/ekspanderbartInnhold.less
