+++ src/Legemeldtsykefravarpanel/Lesmeriawebpanel/InfoToggler/infoToggler.less
@@ -1,4 +1,4 @@
-@import '../../variables.less';
+@import '../../../variables.less';
 
 .infoToggler {
     border: none;
